JyG-Backend/requirements.txt
django>=5.2
djangorestframework>=3.16
django-cors-headers>=4.9
